# openapi.yaml
openapi: 3.0.0
paths:
...
    /movies:
        get:
            description: Returns a paginated, filtered list of movies.
            operationId: MovieController_getAll
            parameters:
                - name: category
                  required: false
                  in: query
                  schema:
                      example: now
                      type: string
                - name: random
                  required: false
                  in: query
                  schema:
                      example: true
                      type: boolean
                - name: limit
                  required: false
                  in: query
                  schema:
                      default: 10
                      example: 12
                      type: number
                - name: page
                  required: false
                  in: query
                  schema:
                      default: 1
                      example: 1
                      type: number
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PaginatedMoviesResponse"
            summary: Get movies
            tags:
                - Movie
        post:
            description: Creates a new movie with a poster and optional banner. Admin only.
            operationId: MovieController_create
            parameters: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - title
                                - description
                                - duration
                                - poster
                            properties:
                                title:
                                    type: string
                                    example: Interstellar
                                slug:
                                    type: string
                                    example: interstellar
                                description:
                                    type: string
                                    example: A team of explorers travel through a wormhole in space.
                                duration:
                                    type: integer
                                    example: 169
                                releaseYear:
                                    type: integer
                                    example: 2014
                                releaseDate:
                                    type: string
                                    format: date
                                    example: "2014-11-07"
                                ratingAge:
                                    type: integer
                                    example: 12
                                country:
                                    type: string
                                    example: USA
                                categoryId:
                                    type: string
                                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                                poster:
                                    type: string
                                    format: binary
                                banner:
                                    type: string
                                    format: binary
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CreateMovieResponse"
            security:
                - bearer: []
            summary: Create movie
            tags:
                - Movie
    /movies/id/{id}:
        get:
            description: Returns a single movie by its id, for admin editing.
            operationId: MovieController_getById
            parameters:
                - name: id
                  required: true
                  in: path
                  schema:
                      type: string
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetMovieResponse"
                "404":
                    description: ""
            security:
                - bearer: []
            summary: Get movie by id
            tags:
                - Movie
    /movies/{slug}:
        get:
            description: Returns a single movie by its slug.
            operationId: MovieController_getBySlug
            parameters:
                - name: slug
                  required: true
                  in: path
                  schema:
                      type: string
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetMovieResponse"
                "404":
                    description: ""
            summary: Get movie by slug
            tags:
                - Movie
    /movies/{id}:
        patch:
            description: >-
                Updates a movie. Only provided fields are changed. Replacing
                poster/banner deletes the old file. Admin only.
            operationId: MovieController_update
            parameters:
                - name: id
                  required: true
                  in: path
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    example: Interstellar
                                slug:
                                    type: string
                                    example: interstellar
                                description:
                                    type: string
                                    example: A team of explorers travel through a wormhole in space.
                                duration:
                                    type: integer
                                    example: 169
                                releaseYear:
                                    type: integer
                                    example: 2014
                                releaseDate:
                                    type: string
                                    format: date
                                    example: "2014-11-07"
                                ratingAge:
                                    type: integer
                                    example: 12
                                country:
                                    type: string
                                    example: USA
                                categoryId:
                                    type: string
                                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                                poster:
                                    type: string
                                    format: binary
                                banner:
                                    type: string
                                    format: binary
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UpdateMovieResponse"
                "404":
                    description: ""
            security:
                - bearer: []
            summary: Update movie
            tags:
                - Movie
        delete:
            description: >-
                Deletes a movie and its images. Fails if it has upcoming screenings.
                Admin only.
            operationId: MovieController_delete
            parameters:
                - name: id
                  required: true
                  in: path
                  schema:
                      type: string
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DeleteMovieResponse"
                "404":
                    description: ""
            security:
                - bearer: []
            summary: Delete movie
            tags:
                - Movie
...
    /screenings/movie/{id}:
        get:
            description: >-
                Returns a paginated list of screenings for a given movie, optionally
                filtered by date.
            operationId: ScreeningController_getByMovie
            parameters:
                - name: id
                  required: true
                  in: path
                  schema:
                      type: string
                - name: date
                  required: false
                  in: query
                  schema:
                      example: "2025-11-08"
                      type: string
                - name: limit
                  required: false
                  in: query
                  schema:
                      default: 20
                      example: 20
                      type: number
                - name: page
                  required: false
                  in: query
                  schema:
                      default: 1
                      example: 1
                      type: number
            responses:
                "200":
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PaginatedScreeningsResponse"
            summary: Get screenings by movie
            tags:
                - Screening
...
info:
    title: Cinema Platform API
    description: API Gateway for Cinema Platform microservices
    version: "1.0"
    contact: {}
...
servers: []
components:
    securitySchemes:
        bearer:
            scheme: bearer
            bearerFormat: JWT
            type: http
    schemas:
...
        GetMoviesResponse:
            type: object
            properties:
                id:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                title:
                    type: string
                    example: Interstellar
                slug:
                    type: string
                    example: interstellar
                poster:
                    type: string
                    example: posters/e0db4163fc2aab147599c750906396d3.jpg
                ratingAge:
                    type: number
                    example: 12
                releaseDate:
                    type: string
                    example: "2014-11-07T00:00:00.000Z"
            required:
                - id
                - title
                - poster
        PaginatedMoviesResponse:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: "#/components/schemas/GetMoviesResponse"
                page:
                    type: number
                    example: 1
                limit:
                    type: number
                    example: 20
                total:
                    type: number
                    example: 137
                totalPages:
                    type: number
                    example: 7
            required:
                - data
                - page
                - limit
                - total
                - totalPages
        GetMovieResponse:
            type: object
            properties:
                id:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                title:
                    type: string
                    example: Interstellar
                slug:
                    type: string
                    example: interstellar
                description:
                    type: string
                    example: A team of explorers travel through a wormhole in space.
                poster:
                    type: string
                    example: posters/e0db4163fc2aab147599c750906396d3.jpg
                banner:
                    type: string
                    example: banners/e0db4163fc2aab147599c750906396d3.jpg
                duration:
                    type: number
                    example: 169
                ratingAge:
                    type: number
                    example: 12
                country:
                    type: string
                    example: USA
                releaseDate:
                    type: string
                    example: "2014-11-07T00:00:00.000Z"
                releaseYear:
                    type: number
                    example: 2014
                categoryId:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
            required:
                - id
                - title
                - description
                - poster
                - duration
        CreateMovieResponse:
            type: object
            properties:
                id:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                title:
                    type: string
                    example: Interstellar
                slug:
                    type: string
                    example: interstellar
                description:
                    type: string
                    example: A team of explorers travel through a wormhole in space.
                poster:
                    type: string
                    example: posters/e0db4163fc2aab147599c750906396d3.jpg
                banner:
                    type: string
                    example: banners/e0db4163fc2aab147599c750906396d3.jpg
                duration:
                    type: number
                    example: 169
                ratingAge:
                    type: number
                    example: 12
                country:
                    type: string
                    example: USA
                releaseDate:
                    type: string
                    example: "2014-11-07T00:00:00.000Z"
                releaseYear:
                    type: number
                    example: 2014
                categoryId:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
            required:
                - id
                - title
                - description
                - poster
                - duration
        UpdateMovieResponse:
            type: object
            properties:
                id:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                title:
                    type: string
                    example: Interstellar
                slug:
                    type: string
                    example: interstellar
                description:
                    type: string
                    example: A team of explorers travel through a wormhole in space.
                poster:
                    type: string
                    example: posters/e0db4163fc2aab147599c750906396d3.jpg
                banner:
                    type: string
                    example: banners/e0db4163fc2aab147599c750906396d3.jpg
                duration:
                    type: number
                    example: 169
                ratingAge:
                    type: number
                    example: 12
                country:
                    type: string
                    example: USA
                releaseDate:
                    type: string
                    example: "2014-11-07T00:00:00.000Z"
                releaseYear:
                    type: number
                    example: 2014
                categoryId:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
            required:
                - id
                - title
                - description
                - poster
                - duration
        DeleteMovieResponse:
            type: object
            properties:
                ok:
                    type: boolean
                    example: true
            required:
                - ok
...
        ScreeningTheaterInfo:
            type: object
            properties:
                id:
                    type: string
                    example: 81lI6j9Ctva6e7oCPpelc
                name:
                    type: string
                    example: Cinema City
                address:
                    type: string
                    example: 123 Main St, Kyiv
            required:
                - id
                - name
                - address
        ScreeningHallInfo:
            type: object
            properties:
                id:
                    type: string
                    example: tZwcb0sTC-0eiRZHm59r9
                name:
                    type: string
                    example: Hall 1
            required:
                - id
                - name
        ScreeningMovieInfo:
            type: object
            properties:
                id:
                    type: string
                    example: 026f20c4-4826-4bea-b5e0-72ce8dda53d4
                title:
                    type: string
                    example: Interstellar
                slug:
                    type: string
                    example: interstellar
                duration:
                    type: number
                    example: 169
                banner:
                    type: string
                    example: banners/e0db4163fc2aab147599c750906396d3.jpg
                poster:
                    type: string
                    example: posters/e0db4163fc2aab147599c750906396d3.jpg
            required:
                - id
                - title
                - slug
                - duration
                - banner
                - poster
        ScreeningSeatType:
            type: object
            properties:
                type:
                    type: string
                    example: vip
                price:
                    type: number
                    example: 250
            required:
                - type
                - price
        GetScreeningsResponse:
            type: object
            properties:
                id:
                    type: string
                    example: X6KBMTnCDkbrPN_I0YXG8
                startAt:
                    type: string
                    example: "2026-12-21 16:30:00.000"
                endAt:
                    type: string
                    example: "2026-12-21 19:30:00.000"
                theater:
                    $ref: "#/components/schemas/ScreeningTheaterInfo"
                hall:
                    $ref: "#/components/schemas/ScreeningHallInfo"
                movie:
                    $ref: "#/components/schemas/ScreeningMovieInfo"
                seatTypes:
                    type: array
                    items:
                        $ref: "#/components/schemas/ScreeningSeatType"
            required:
                - id
                - startAt
                - endAt
                - theater
                - hall
                - movie
                - seatTypes
        PaginatedScreeningsResponse:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: "#/components/schemas/GetScreeningsResponse"
                page:
                    type: number
                    example: 1
                limit:
                    type: number
                    example: 20
                total:
                    type: number
                    example: 84
                totalPages:
                    type: number
                    example: 5
            required:
                - data
                - page
                - limit
                - total
                - totalPages
...